Created a Clone method in the BrainParameters, created a new BrainParameters file, made a CumSum method
