+++ UnitySDK/Assets/ML-Agents/Editor/BrainEditor.cs
@@ -14,26 +14,7 @@ namespace MLAgents
     /// </summary>
     [CustomEditor(typeof(Brain))]
     public class BrainEditor : Editor
-    {
-        /// <summary>
-        /// DeepCopy of BrainParameters
-        /// </summary>
-        /// <param name="src">The BrainParameters that will be copied</param>
-        /// <param name="dst">The BrainParameters that will be overwritten</param>
-        private static void DeepCopyBrainParameters(
-            BrainParameters src, ref BrainParameters dst)
-        {
-            dst = new BrainParameters()
-            {
-                vectorObservationSize = src.vectorObservationSize,
-                numStackedVectorObservations = src.numStackedVectorObservations,
-                vectorActionSize = (int[]) src.vectorActionSize.Clone(),
-                cameraResolutions = (resolution[]) src.cameraResolutions.Clone(),
-                vectorActionDescriptions = (string[]) src.vectorActionDescriptions.Clone(),
-                vectorActionSpaceType = src.vectorActionSpaceType
-            };
-        }
-        
+    {        
         public override void OnInspectorGUI()
         {
             var brain = (Brain) target;
@@ -41,7 +22,7 @@ namespace MLAgents
                 "Copy Brain Parameters from : ", null, typeof(Brain), false) as Brain;
             if (brainToCopy != null)
             {
-                DeepCopyBrainParameters(brainToCopy.brainParameters, ref brain.brainParameters);
+                brain.brainParameters = brainToCopy.brainParameters.Clone();
             }
             var serializedBrain = serializedObject;
             serializedBrain.Update();
