+++ UnitySDK/Assets/ML-Agents/Editor/Tests/MLAgentsEditModeTest.cs
@@ -253,7 +253,7 @@ namespace MLAgents.Tests
             agent2.agentParameters.onDemandDecision = true;
             // agent2 will request decisions only when RequestDecision is called
             brain.brainParameters.vectorObservationSize = 0;
-            brain.brainParameters.cameraResolutions = new resolution[0];
+            brain.brainParameters.cameraResolutions = new Resolution[0];
             agent1.GiveBrain(brain);
             agent2.GiveBrain(brain);
 
@@ -391,7 +391,7 @@ namespace MLAgents.Tests
             agent2.agentParameters.onDemandDecision = true;
             // agent2 will request decisions only when RequestDecision is called
             brain.brainParameters.vectorObservationSize = 0;
-            brain.brainParameters.cameraResolutions = new resolution[0];
+            brain.brainParameters.cameraResolutions = new Resolution[0];
             agent1.GiveBrain(brain);
             agent2.GiveBrain(brain);
 
@@ -576,7 +576,7 @@ namespace MLAgents.Tests
             agent1.agentParameters.maxStep = 20;
             agent2.agentParameters.maxStep = 30;
             brain.brainParameters.vectorObservationSize = 0;
-            brain.brainParameters.cameraResolutions = new resolution[0];
+            brain.brainParameters.cameraResolutions = new Resolution[0];
             agent1.GiveBrain(brain);
             agent2.GiveBrain(brain);
 
@@ -706,7 +706,7 @@ namespace MLAgents.Tests
             agent1.agentParameters.resetOnDone = false;
             agent2.agentParameters.resetOnDone = false;
             brain.brainParameters.vectorObservationSize = 0;
-            brain.brainParameters.cameraResolutions = new resolution[0];
+            brain.brainParameters.cameraResolutions = new Resolution[0];
             agent1.GiveBrain(brain);
             agent2.GiveBrain(brain);
 
@@ -788,7 +788,7 @@ namespace MLAgents.Tests
             // agent2 will request decisions only when RequestDecision is called
             agent1.agentParameters.maxStep = 20;
             brain.brainParameters.vectorObservationSize = 0;
-            brain.brainParameters.cameraResolutions = new resolution[0];
+            brain.brainParameters.cameraResolutions = new Resolution[0];
             agent1.GiveBrain(brain);
             agent2.GiveBrain(brain);
