+++ UnitySDK/Assets/ML-Agents/Scripts/Brain.cs
@@ -1,109 +1,8 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
-using System.Linq;
 
 namespace MLAgents
 {
-
-    public enum SpaceType
-    {
-        discrete,
-        continuous
-    };
-
-
-/** Only need to be modified in the brain's inpector.
- * Defines what is the resolution of the camera
-*/
-    [System.Serializable]
-    public struct resolution
-    {
-        public int width;
-
-        /**< \brief The width of the observation in pixels */
-        public int height;
-
-        /**< \brief The height of the observation in pixels */
-        public bool blackAndWhite;
-        /**< \brief If true, the image will be in black and white. 
-         * If false, it will be in colors RGB */
-    }
-
-/** Should be modified via the Editor Inspector.
- * Defines brain-specific parameters
-*/
-    [System.Serializable]
-    public class BrainParameters
-    {
-        public int vectorObservationSize = 1;
-        /**< \brief If continuous : The length of the float vector that represents 
-         * the state
-         * <br> If discrete : The number of possible values the state can take*/
-
-        [Range(1, 50)] public int numStackedVectorObservations = 1;
-
-        public int[] vectorActionSize = new int[1]{1};
-        /**< \brief If continuous : The length of the float vector that represents
-         * the action
-         * <br> If discrete : The number of possible values the action can take*/
-
-        public resolution[] cameraResolutions;
-        /**<\brief  The list of observation resolutions for the brain */
-
-        public string[] vectorActionDescriptions;
-        /**< \brief The list of strings describing what the actions correpond to */
-
-        public SpaceType vectorActionSpaceType = SpaceType.discrete;
-        /**< \brief Defines if the action is discrete or continuous */
-        
-        /// <summary>
-        /// Converts a Brain into to a Protobuff BrainInfoProto so it can be sent
-        /// </summary>
-        /// <returns>The BrainInfoProto generated.</returns>
-        /// <param name="name">The name of the brain.</param>
-        /// <param name="isTraining">Whether or not the Brain is training.</param>
-        public CommunicatorObjects.BrainParametersProto 
-            ToProto(string name, bool isTraining)
-        {
-            var brainParametersProto = new CommunicatorObjects.BrainParametersProto
-            {
-                VectorObservationSize = vectorObservationSize,
-                NumStackedVectorObservations = numStackedVectorObservations,
-                VectorActionSize = {vectorActionSize},
-                VectorActionSpaceType =
-                    (CommunicatorObjects.SpaceTypeProto)vectorActionSpaceType,
-                BrainName = name,
-                IsTraining = isTraining
-            };
-            brainParametersProto.VectorActionDescriptions.AddRange(vectorActionDescriptions);
-            foreach (resolution res in cameraResolutions)
-            {
-                brainParametersProto.CameraResolutions.Add(
-                    new CommunicatorObjects.ResolutionProto
-                    {
-                        Width = res.width,
-                        Height = res.height,
-                        GrayScale = res.blackAndWhite
-                    });
-            }
-            return brainParametersProto;
-        }
-
-        public BrainParameters()
-        {
-            
-        }
-
-        public BrainParameters(CommunicatorObjects.BrainParametersProto brainParametersProto)
-        {
-            vectorObservationSize = brainParametersProto.VectorObservationSize;
-            numStackedVectorObservations = brainParametersProto.NumStackedVectorObservations;
-            vectorActionSize = brainParametersProto.VectorActionSize.ToArray();
-            vectorActionDescriptions = brainParametersProto.VectorActionDescriptions.ToArray();
-            vectorActionSpaceType = (SpaceType)brainParametersProto.VectorActionSpaceType;
-        }
-    }
-
     /// <summary>
     /// Brain receive data from Agents through calls to SendState. The brain then updates the
     /// actions of the agents at each FixedUpdate.
